Changed equdistant conic to be a single class using the old gctp style for setting the mode in the parameter list.
fixed polyconic, now calling init functions

+++ GCTP_CPP/util.cpp
@@ -169,7 +169,7 @@ printf("Latitude failed to converge after 15 iterations -- PHI3Z-CONV");
 return(3);
 }
 
-double Util::phi4z(double eccent, double e0, double e1, double e2, double e3, double a, double b, double* c, double* phi) {
+long Util::phi4z(double eccent, double e0, double e1, double e2, double e3, double a, double b, double* c, double* phi) {
 
 double sinphi;
 double sin2ph;
